+++ pkg/mtlskeystore/mtlskeystore.go
@@ -1,6 +1,7 @@
 package mtlskeystore
 
 import (
+	"context"
 	"errors"
 
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v10/pkg/session"
@@ -13,7 +14,30 @@ var (
 
 type (
 	// MTLSKeystore is the interface for the mTLS Keystore API.
-	MTLSKeystore interface{}
+	MTLSKeystore interface {
+
+		// Client certificates
+
+		// ListClientCertificates lists client certificates under the account.
+		//
+		// See: https://techdocs.akamai.com/mtls-origin-keystore/reference/get-client-certs
+		ListClientCertificates(ctx context.Context) (*ListClientCertificatesResponse, error)
+
+		// GetClientCertificate gets details of a client certificate.
+		//
+		// See: https://techdocs.akamai.com/mtls-origin-keystore/reference/get-client-cert
+		GetClientCertificate(ctx context.Context, params GetClientCertificateRequest) (*GetClientCertificateResponse, error)
+
+		// CreateClientCertificate creates a client certificate with the provided name.
+		//
+		// See: https://techdocs.akamai.com/mtls-origin-keystore/reference/post-client-cert
+		CreateClientCertificate(ctx context.Context, params CreateClientCertificateRequest) (*CreateClientCertificateResponse, error)
+
+		// PatchClientCertificate updates the client certificate's name or notification emails.
+		//
+		// See: https://techdocs.akamai.com/mtls-origin-keystore/reference/patch-client-cert
+		PatchClientCertificate(ctx context.Context, params PatchClientCertificateRequest) error
+	}
 
 	mtlskeystore struct {
 		session.Session